test(observability): consolidate forwarder mock collector + typo fix
Addresses two Bridgebuilder F-findings from the local self-review pass
on PR #1618 head a3c37fb0:

F7 (LOW, maintainability) — `CapturingTraceService`, the `TraceService`
impl, the spawn helper, and the readiness probe were duplicated across
all three `otel_forwarder_*` integration tests (~60 lines each, three
copies). When the `SdkForwarderState` shape changed in earlier rounds
the three helpers drifted out of lockstep until the partial-batch test
caught it. Extracted to `engine/tests/common/forwarder_mock.rs` with
two variants:

- `spawn_body_capturing_collector` — used by the resource-preservation
  and partial-batch tests (both need the `ExportTraceServiceRequest`
  body).
- `spawn_metadata_capturing_collector` — used by the headers test
  (needs `tonic::MetadataMap` instead of body).

Both share the same readiness-probe contract (bounded
`TcpStream::connect` retry; see file-level doc-comment for the
rationale that replaced the 50ms-sleep heuristic earlier in the
review cycle). Registered in `engine/tests/common/mod.rs`
alongside the existing `http_helpers` / `queue_helpers` modules.

Net effect: tests/ shrinks by ~210 lines (210 deletions vs 32
insertions, after the new helper module is counted).

F6 (NIT, spelling) — "preceeding" → "preceding" in the comment block
on the attribute-count assertion in `otel_forwarder_resource_test.rs`.

Deferred F-findings (with rationale): F1 (OnceLock fragility) was
already addressed in round-3 via the doc-comment on
`SDK_SPAN_FORWARDER`; BB couldn't see otel.rs in the truncated diff.
F2/F3 (server task panics / readiness probe noise) — `.expect()` was
added in round-3 and the noise is acceptable per BB's own confidence
(0.4/0.6). F5 (dep placement) — `tonic` IS used in production code
(`Channel`, `Endpoint`, `MetadataKey/Value`, `ClientTlsConfig`); the
`[dependencies]` placement is correct. F8 (truncated otel.rs review
scope) — adversarial cross-model dissent (`gpt-5.3-codex`) reviewed
the full diff and returned `findings: []` after the round-4
spec-length tightening.

Sweep: 1542 lib tests passed, 3 forwarder integration tests passed
(all three now via the shared helper module), cargo fmt + clippy
clean on touched files.

Ridden with [Loa](https://github.com/0xHoneyJar/loa)
