+++ engine/tests/common/mod.rs
@@ -1,3 +1,4 @@
+pub mod forwarder_mock;
 pub mod http_helpers;
 pub mod queue_helpers;
